+++ src/data/prod/README.MD
@@ -1,12 +1,48 @@
-# Creating data
-Application data that is not user-generated must be created (or potentially updated), and `src/data/prod/upsert.ts` is a script to do that.
+# Manipulating Data
+Application data that is not user-generated must be created (or potentially updated), and thankfully we have a couple scripts to do just that.
+
+## Prerequisites
+Local DB connections need to be established (via ssh tunnels or whatever you want) before running.
+
+## Creating new user data
 From csv data with formats defined [here](https://docs.google.com/spreadsheets/d/1H7Q5sPkcGqgdZdqzJpH2f-PluQgEfLZ6Dxy4xmIQTZ4/edit#gid=664790308), the script optionally creates
 - Organizations and their funding spaces
 - Sites
 - Users and their permissions
 - Reporting periods
 
+### Usage
+```
+$ yarn ts-node src/data/prod/create.ts --config [CONFIG_FILE_PATH]
+```
+where:
+- `CONFIG_FILE_PATH` is an absolute path to the config file
+
+example:
+```
+$ cp src/data/prod/config.example.json ./config.json
+$ vim config.json // edit config to contain real values!
+$ yarn ts-node src/data/prod/create.ts --config `pwd`/config.json
+```
+
 Alternatively, script can generate and send invite emails (through the winged-keys app) to users so they can set passwords and use the tool!
+
+## Updating existing user data
+There are times where we may need to update a user's email within WingedKeys, and will need to propagate that update over to Fawkes.  The `update` script handles just that, porting the state of all users in WingedKeys and applying that state to all corresponding users in Fawkes.
+
+### Usage
+```
+$ yarn ts-node src/data/prod/update.ts --config [CONFIG_FILE_PATH]
+```
+where:
+- `CONFIG_FILE_PATH` is an absolute path to the config file
+
+example:
+```
+$ cp src/data/prod/config.example.json ./config.json
+$ vim config.json // edit config to contain real values!
+$ yarn ts-node src/data/prod/update.ts --config `pwd`/config.json
+```
 
 ## Configuration
 Configuration is supplied as a JSON file like:
@@ -45,27 +81,10 @@ Configuration is supplied as a JSON file like:
 **NOTE**: instead of passing app db overrides, you can just update your local ormconfig but DON'T FORGET TO CHANGE IT BACK.
 **NOTE**: At least one file param must be supplied.
 
-## Usage
-```
-$ yarn ts-node src/data/prod/create.ts --config [CONFIG_FILE_PATH] [--invite]
-```
-where:
-- `CONFIG_FILE_PATH` is an absolute path to the config file
-- `--invite` flag means all data creation is skipped, and invite emails are generated and sent to all users in `userFile`
-
-example:
-```
-$ cp src/data/prod/config.example.json ./config.json
-$ vim config.json // edit config to contain real values!
-$ yarn ts-node src/data/prod/create.ts --config `pwd`/config.json
-$ yarn ts-node src/data/prod/create.ts --config `pwd`/config.json --invite
-```
-
 ## Docker
 If you're running this is in docker (for instance, on a node image), you'll need to install a few dependencies to get puppeteer working.
 Run the `apt-get` command they've got in their puppeteer-in-docker [example Dockerfile](https://github.com/puppeteer/puppeteer/blob/main/docs/troubleshooting.md#running-puppeteer-in-docker)
 
 ## Gotchas
-- Local DB connections need to be established (via ssh tunnels or whatever you want) before running
 - Any changes to the template columns (not titles, just order) will require changes to the corresponding row class defined in the various data creation utils
 - Org/site names cannot have commas (",") in them because this will mess with parsing. So just take them out, sorry.
