+++ src/data/prod/users/config.ts
@@ -1,0 +1,12 @@
+export type SiteConnectionOpts = {
+  url: string;
+  user: string;
+  password: string;
+};
+
+export type DBConnectionOpts = {
+  server: string;
+  port: number;
+  user: string;
+  password: string;
+};
